Updated openstack/openstack Project: openstack/neutron  8627fad60724df09ce2f58d7b1ae2932702036f3
Update related router when subnetpool change scope

When the address scope of a subnetpool changes, the scope of related
subnet will change too. If this subnet has been connected to a router,
the related routing information needs to be updated.
This patch will fire the router update to correlative routers. The
routers will get all things updated in the update action.

Change-Id: Ia2eb484eb949a8d42fc909291de7be07b002adda
Partially-Implements: blueprint address-scopes