Updated openstack/openstack Project: openstack/manila  f40ca30007b1b3b51e4b365f9cf6566451f1e56b
TrivialFix: Remove deprecated option 'DEFAULT/verbose'

Option "verbose" from group "DEFAULT" is deprecated for removal.
Its value may be silently ignored in the future.

If this option is not set explicitly, there is no such warning.
Furthermore, the default value of verbose is true, so there is no
need to set this value in config files.

And we use oslo_config.generator to generate a sample file.
We'd better fix the doc not encourage user to set verbose=True.

Change-Id: Ie9a68e511425d34cbce04e416f258db3064d0479