Updated openstack/openstack Project: openstack-infra/system-config  857b22519af1025b450b6d0f3de6d747a073f429
Remove restricted and multiverse

The packages in restricted and multiverse are either non-free or
unsupported. While they're usually useful for desktop, they aren't
something that OpenStack should grow a dependency on.

Also remove non-free and contrib from debian. We don't do anything with
debian yet - but if we did, I'm pretty sure we would not want these.

Note: The existing restricted and multiverse dirs should be purged,
although it's not urgent to do so. 'reprepro --delete clearvanished'

Change-Id: I1d98ab1dea52598d01e3ce3d9a24306280a190fe