Updated openstack/openstack Project: openstack/keystone  8fab958346ba974d63381b7a9f7c91493e850737
Set deprecated_reason on deprecated config options

version 3.2.0 of oslo.config supports a deprecated_reason to
help explain to deployers why an option will soon be removed.
This patch aims to provide a good explanation to the currently
deprecated options.

Note: The eventlet related options that are deprecated for removal
were not changed since they are slated for removal in Mitaka.

Change-Id: I1befef16e1266b9e08c83ba5beb498f669d066ae