Updated openstack/openstack Project: openstack/swift  98589640ced22c38ac97438d0d63e85e332881d6
Monkeypatch everything the same

There's three spots in the code where eventlet monkeypatches
things. In one place, it was monkeypatching socket and thread; in the
other two, only socket. This commit monkeypatches socket and thread
everywhere.

Hopefully, this lets us avoid code that, for example, works in the
object server but fails in the object updater.

Change-Id: Ia30d069c5ac89f4bfbe52233d03d96738144b614