Updated openstack/openstack Project: openstack/keystone  9202748ccdbc93ab5e675a8a3a181e7a678a53a6
Add in TRACE logging for the manager

Add a metaclass that will wrap all manager methods for trace
logging. Trace logging will encompass Call of the method,
args, kwargs, runtime, and return value/exception raised.

TRACE logging should not be enabled except in extreme debug
circumstances. When TRACE log level is not set, the logger will
not construct the messages as it is potentially expensive.

Change-Id: Idd2656e00e128914e00a9fdec410755697df33db