Updated openstack/openstack Project: openstack/nova  527ba87ff051f8c4a02f6f9d46b0e1f9ac6cd999
Allow all api_samples tests to be run individually

oslo.config is very specific about what is allowed to be overridden in
our flags method. When run as a whole, the functional tests import
enough stuff so that all the tests pass. But if you attempt to run
just the versions tests, they will fail with an import error. This is
because the api_samples_base references osapi_compute_link_prefix and
osapi_compute_extension in all cases.

Adding this to the base class makes these tests runnable on their own.

Change-Id: I6e9800fdfb7a3c8495de71220bfec0597e45da82