Updated openstack/openstack Project: openstack-infra/project-config  7d76e46a832ab79d92d99d650cd87c78df71a17b
Revert "Enable voting for Puppet OpenStack Integration jobs"

We're having CI issues at this time, and need to land some patches that
won't pass integration jobs.

We need to patch puppet-nova and puppet-gnocchi required to pass
integration jobs but we'll also need to patch
puppet-openstack-integration to make all them pass.
Because of this chicken & egg problem, we want to disable voting for
today and eventually tomorrow, until we sort this out.

This reverts commit e1cee88bea0a38cb85168e009a6ebdc2aaa747b7.

Change-Id: Iab40d374d4f37350519dfc961e5f0a1b14847abf