Updated openstack/openstack Project: openstack/nova  674531df15d30faea08a58421160816c56b4bfba
Add new test_rebuild_instance_with_volume to cells exclude list

This commit adds a newly added tempest test to the cells exclude list.
The test is always failing on cells and it's because it's doing
operations that don't work with cells turned on. So lets exclude it
for now. Ideally all this skip logic will be in the tempest config and
we don't have to do this anymore.

Closes-Bug: #1543253
Change-Id: Ic9db51a41f95b0d18f97745a0da7e99fdfa21e51