Updated openstack/openstack Project: openstack/keystone  8eacd48eded91c80d43070c3c27fb517447b9bce
Support multiple URLs for LDAP server

python-ldap calls out to openldap which can handle multiple URLs for
ldap servers (for the purpose of high availability). openldap expects
these urls to be separated by a comma or whitespace.

Change the help text to specify a comma separated list of URLs is
allowed.

Change-Id: I523dcfc1701a6f7c725c4aa11482bfc15a3515a5
Closes-Bug: #1500631