Updated openstack/openstack Project: openstack/requirements  612cdd78dc6b4c99dda8b65b8b2204ec4d61acc2
Bump up eventlet to at least 0.18.1

The eventlet version starting with 0.18.0 has a fix
that allows for running the evenlet backdoor off of
a local unix domain socket (which can be made more secure
than a random local port).

The fix is at:

https://github.com/eventlet/eventlet/commit/24d283cfd

Said functionality would be nice to have so that
oslo_service can provide that backdoor over a more secureable
channel (one that does not lose the concept of unix
users/groups).

Change-Id: Ia2385879e09991102f8f305ec41dbb651a4374de