Updated openstack/openstack Project: openstack-infra/system-config  26bb470549443e55726f85ff5161a24ba7e5eb22
Don't write k5start banner to file list

k5start prints:

  "Kerberos initialization for service/reprepro@OPENSTACK.ORG"

This then gets redirected into the list of files to remove on the next
run - which is an error, since that's not a filename. Put the
redirection inside of a bash command so that only the data gets
redirected.

Change-Id: Icae10763b4bce78c7b674715826547c228980dd6