Updated openstack/openstack Project: openstack/neutron  4e42dc41d6235e35ea0ae8fe7c8bd59474065ff8
Make add_tap_interface resillient to removal

This patch makes add_tap_interface safe to race conditions
where the interface is removed in the middle of processing
by catching exceptions and checking to see if the interface
still exists. If it no longer exists it assumes the exception
was caused by the missing interface and returns False as it
would if the interface did not exist to begin with.

Change-Id: Ie0d89fc2584490b6985aee66da70bae027a130ed
Closes-bug: #1542972