Updated openstack/openstack Project: openstack/horizon  64001b67ae3121d6e477450166f0a3f03ec88963
Fix Create Image angularjs code

This code intended to allow pre-filling of the form values through
Django's handling of GET parameters, but it did not do so as the
code was incorrect. This was not previously noticed as the angularjs
code wasn't actually being executed. Once it was, it broke. Ironic,
I know.

Change-Id: I8d641de9246fd4f43c96bf85d47bb648f4401def
Closes-Bug: 1503396