Updated openstack/openstack Project: openstack/ironic  5a1c117ab63f11bfff0c48e50e92dbaf6fe06873
Log warning if ipmi_username/ipmi_password missing

Even though an empty ipmi_username and/or ipmi_password is valid, it is
often an error in configuration. Log a warning if either value is
missing.

Add unit tests showing that an unspecified username or an empty string
username are treated the same and will cause no user to be specified.

Add unit tests showing that an unspecified password or an empty string
password are treated the same and will cause a NULL (\0) password to be
used.

Change-Id: I384e210da3216fb9335c4c782d69e606026d4b97