Updated openstack/openstack Project: openstack/tempest  04d8c9d9791bc464c68ffe87471f704d09748075
Update .mailmap

This commit updates and sorts the .mailmap file for top 30th
contributors. I picked up the latest author names and emails from the
commit log.
The .mailmap feature is used to coalesce together commits by the same
person in the shortlog, where their name and/or email address was
spelled differently[1].

[1] https://git-scm.com/docs/git-shortlog

Change-Id: Ibc28a0b4f5839df9e6cd7df156ed818300312acd