Updated openstack/openstack Project: openstack-infra/project-config  7b24ceecfddd9ec37148b412f0354d02bf356393
Update configure_mirror to use per-region APT mirrors

Now that we have per-region apt mirrors for ubuntu, we should use them.
As part of doing that we need to disable looking for i386 packages since
we do not mirror nor use those - and we need to tell APT to not check
the GPG signatures of our mirror archive since we're not signing it.

We don't need multiverse or restricted in the gate, and they also
contain non-free software. Remove them from our configuration.

Change-Id: I26f5fc5cec224cf4f4ea361e0f71cecbf320fbb0