Updated openstack/openstack Project: openstack/nova  d754a830861fb55b047e7b4d43ba7f485fc120dd
Replace eventlet-based raw socket client with requests

There's no discernable reason why we need to use a raw
eventlet-based client TestWSGIServerWithSSL to confirm that a
wsgi server is listening correctly and returning the expected
response, especially since using eventlet leads to unreliable
i/o on the socket unless there is an explicit yield before
asserting the value of the response.

So requests is used to query the servers and verify the response.

Change-Id: I00c1c1e93eb7c66c3370434db4e3e6c00087f8aa
Closes-Bug: #1543766