Updated openstack/openstack Project: openstack-infra/release-tools  bb052de702bcc19983c989747d1186f78410ab48
process pypi flag for release announcements

When tagging a release, look for a flag telling us whether to include
the PyPI link in release announcements and save the information to the
tag message.

When announcing a release, look for the flag and only include the PyPI
link when it is set to yes.

Change-Id: I8e0a2517d2a8c19381d2d57600da92c954775e5e
Depends-On: I491fb9e6edd57a3ec1c0b85c8dd9e22d5f09e6b2
Signed-off-by: Doug Hellmann <doug@doughellmann.com>