Updated openstack/openstack Project: openstack/cinder  e88f0a486ab4f9c7baaf5d7d3781705221e89146
Zfssaiscsi driver should return target_lun as int

The zfssaiscsi driver in initialize_connection()
returns iscsi_properties['target_lun'] as a string.
Solaris version of openstack expects it to be an int
and fails for the driver. Also other drivers report
it as an int. So this will make the zfssaiscsi driver
more consistent with other drivers.

Change-Id: Ifdbca7a02e2d797af9149cd185c2fa5dce661f66
Closes-bug: #1538582