Updated openstack/openstack Project: openstack/neutron  3c2932eb687bee26450ee5aaae19657aa2ee17cf
L3: enable plugin to decide if subnet is mandatory

Commit 07077bebb69da29994257d061d3a8d7ea9598c3d removed
enforcing that a subnet should be defined on a network. In
the case of IPv6 this is cool, but for V4 it is problematic.

The patch adds in a callback that enables the plugin to
determine validation policy.

The callback will be invoked prior to creating the gateway
port.

Closes-bug: #1543012

Change-Id: Idb7eeb5e0071aa5a2f302fba083504c4582edc1a