Updated openstack/openstack Project: openstack-infra/release-tools  8e6e7ea6d89dcb1b3b4f833fbeac2652ab90c6ad
Do not mention "independent" in announcements

Projects following the independent release model do not use any
release series. However announcements include a line saying
"This release is part of the independent release series", and the
subject line mentions "(independent)".

This is confusing and an artifact of our process that should not
be exposed in announcements, so we should just not mention
"independent" (or any series) in that case.

Change-Id: I9dbc20a13eafa311c03447db005ceefa97f34871