Updated openstack/openstack Project: openstack/neutron  dbb26a97772997c91a7b3670f8e293637b3ed66d
Test helpers to facilitate testing BGP dynamic routing.

This patch includes changes that facilitate creation of subnetpools
and address scopes, as well as changes that make it easier to allocate
subnets from a subnetpool inside unit and API tests. These fixtures
are needed for testing BGP features, but have general utility in
in developing Neutron tests in the future.

Change-Id: I65749dac516e3ff23db50cbb7b832aa3039394e6
Partially-Implements: blueprint bgp-dynamic-routing