Updated openstack/openstack Project: openstack/ironic  0b3da478a24a85df392ad383acea15bb559b80dc
Remove 'zapping' from code

The term 'zapping' has been replaced by 'manual cleaning'. This
removes 'zapping' from the code. Since one of the changes is in
the help for a config option, ironic.conf.sample was updated
(note that it was out of date).

(References to 'zapping' in the documentation is addressed in a
separate patch.)

Change-Id: Icc0303cf26b464b84ba97636b41d9edfa16b8dbd
Partial-Bug: #1526290