Updated openstack/openstack Project: openstack-dev/devstack  1a9e294ff398e5f3c3a0a62621b8872273aeb0f9
Switch tempest verify-config usage to avoid bug

This commit will unblock the world breakage caused by the recent
cliff release. This exposes a hole in our constraints usage in
devstack. We need to fix this bug for real in devstack and cliff,
but that will take a few minutes to an hour or so. So let's just
change the usage in the meantime, we can revert this when things
are fixed for real.

Related-Bug: #1543841

Depends-On: Ic63612dc50e064a3a69b88618e394ba17e083c22
Change-Id: I19e477fa5068474bc3471307732f89adaafb2952