Updated openstack/openstack Project: openstack-infra/project-config  1edeccf1a4dc78b3287791395196b255b460ca42
Rename periodic networking jobs

Rename periodic networking jobs so that they are named as "periodic-X"
instead of "gate-X".

For the master jobs, add explicit branch of master so that the right
branch is used. Note that via devstack-jobs, the ovs jobs are already
defined as periodic jobs.

Remove now unneeded non-voting conditions.

Change-Id: Iac4671f3762a7855c4870cd061b7ce282a077e53