Updated openstack/openstack Project: openstack/neutron  5888f79baf634202d23b02df91b5e973dfa0eab8
stadium: Propose kuryr as an independent project.

This patch adds an additional piece of criteria for Neutron sub-projects.
Projects that interface with Neutron on REST API boundaries only should
probably be separate.  We propose Kuryr be split out based on this criteria.
We also document why Octavia stays for now.

Change-Id: Ic161409f6d1ca2efb623d9c7c2797d158a8094df
Signed-off-by: Russell Bryant <rbryant@redhat.com>