Updated openstack/openstack Project: openstack/nova  f69ee4a8b1cc15f0be2a765895791861096242bf
Remove and deprecate conductor provider_fw_rule_get_all()

This is unused and can be cleaned out. The manager method is left for
backwards compatibility but changed to return an empty list since that
yields the same result as before. The db query always resulted in no
rows being returned since the table contained nothing.

Change-Id: Id5f0406c64f9a6ef9e4220cd6e6d2facbf40a550