Updated openstack/openstack Project: openstack/nova  58dafee0d931e2330e9ee6012d5558e022fc26fe
Get a ReqSpec in evacuate API and pass it to scheduler

Since the RequestSpec object is now persisted in the API layer every time that
we start an instance, we can now fetch it from the API DB and pass it thru the
ComputeTask API to the conductor so it can directly call the scheduler with it.

NOTE(sbauza): Yeah, it's ugly to see the ReqsSpec object being dehydrated to
legacy dicts and then rehydrated by the next call, but I'll modify that in my
next change.

Partially-Implements: blueprint check-destination-on-migrations

Change-Id: I7fe694175bb47f53d281bd62ac200f1c8416682b