Updated openstack/openstack Project: openstack/nova  8615de1bb7afac8ffbd7d9c8f8e7235c49df9b39
Fix fallocate test on newer util-linux

Newer util-linux raises an error when calling fallocate with the -n
option if the target file does not already exist. This is because the
-n option directs it to retain the file's existing size. A
non-existent file does not have an existing size. fallocate in older
releases of util-linux creates a zero-sized file in this case. This
results in _can_fallocate() always returning false, and therefore
never preallocating.

While this may reasonably be argued to be a regression in util-linux,
the -n option doesn't make sense here anyway, so we remove it.

Closes-Bug: #1543181

Change-Id: Ie96fa71e7d2641d30572b8eda5609dd3ca5b6708