Updated openstack/openstack Project: openstack/oslo.messaging  9ae976d8d6c4fd7e04860e52687fb99003269280
[kombu] Implement experimental message compression

Add two new options disabled by default to set the
kombu serializer and compression algorithm:
http://docs.celeryproject.org/projects/kombu/en/latest/reference/kombu.html#message-producer

During the spec discussion, it was apparent that drivers
were doing things differently (ZMQ and pickling for example).
There was concerns around the CPU vs message size tradeoff among
other things:
https://review.openstack.org/#/c/247668/

So in this review, we leave the defaults as it was before, but
add EXPERIMENTAL options, to see how these holds up in
a large stress test and will help test these in the CI as well.

Change-Id: If240774657c2175874c45243db7854ba372982bd