Updated openstack/openstack Project: openstack-infra/elastic-recheck  05bae54bc6001e1da75a483a8685853d92bc4ec0
Remove query for bug 1524418

The error:

"AttributeError: 'LocalManager' object has no attribute 'l3driver'"

Was fixed in oslo.messaging 4.0.0. The generic MessagingTimeout
part of this query is now spiking on something new, so delete this
query so we can file a new bug and track with a new query.

Change-Id: Ie21178d68f84642ba93eca652bfd98778ea90fa8
Related-Bug: #1524418