Updated openstack/openstack Project: openstack/python-swiftclient  84d110c63ecf671377d4b2338060e9b00da44a4f
Accept token and tenant_id for authenticating against KS

Allow swiftclient to authenticate against keystone using tenant
name/id and token only. Without this patch, the password is
required, which may not always be available. Authentication
against keystone is required to get the service catalog,
which includes the endpoints for swift.

Change-Id: I4477af445474c5fa97ff864c4942f1330b59e5d6
Closes-Bug: #1476002