Updated openstack/openstack Project: openstack/nova  f39b2b242de3e5124603b86f585a20b6596a35d1
Add better help text to scheduler options (5)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 5 of 7.

Blueprint centralize-config-options

Change-Id: Iac694747a219ed690f8d75068982df8cbecae28b