Updated openstack/openstack Project: openstack/nova  91010e1b50a24ea8081a2b0771a2be3b8d3a6e92
Add a tool for reserving migration placeholders during release time

This adds a tool to make it easy to reserve placeholder migrations
after we do a release. This is something we have to do each time we
do a release, and it can be somewhat silly in terms of mechanical
work. This adds a script to make it trivially easy.

Change-Id: I34c11f1aee64ad36887779f1c9ca4491a25580ec