Updated openstack/openstack Project: openstack/keystone  5bb5bea3f0662724a3ae27e7fd9e725dfcbbdde2
Remove support for trusts in v2.0

If the UUID token provider is configured, it is possible to use trusts to
authenticate against the version 2.0 API. The fernet token provider doesn't
support trust authentication against the v2.0 API. Trusts were never supported
in version 2.0 and it is recommend that trusts are used exclusively with
version 3.

This commit makes this behavior with trusts and v2.0 consistent regardless of
the configured token provider. It is also a step to getting fernet to be the
default token provider in Keystone.

Change-Id: Ia02850d1aec304eccccb4752c43767272047b74e